gme_trading_system/requirements.txt
flask
supervisor
crewai>=0.28.0
crewai-tools
apscheduler
pandas
gitpython
python-dotenv
streamlit
pyyaml
openai
google-generativeai
langchain-google-genai
finnhub-python
ccxt
sentence-transformers
chromadb
requests
yfinance
ib_insync
websocket-client
supabase
